Fixed #42 and #45. Implemented similar fixes for save profiles.
