+++ src/main/resources/version.properties
@@ -1,2 +1,2 @@
-#Thu Jan 09 19:12:51 EST 2025
+#Thu Jan 09 19:34:09 EST 2025
 version=0.4.6
